added new logic

+++ src/components/DeleteButton.js
@@ -2,24 +2,26 @@ import { toast } from 'react-toastify';
 import { FaTrash } from 'react-icons/fa';
 
 import Button from './Button';
-import { DELETE_STORY } from 'context/types';
 import { deleteStory } from 'services/storyService';
-import { useGlobalContext } from 'context/Context';
+import { DELETE_STORY } from 'context/story/StoryTypes';
+import { useGlobalContext } from 'context/story/StoryContext';
 
 const DeleteButton = ({ id }) => {
   const { dispatch } = useGlobalContext();
 
   const handleDelete = async (id) => {
-    try {
-      dispatch({
-        type: DELETE_STORY,
-        payload: id,
-      });
-      await deleteStory(id);
-      window.location.reload();
-    } catch (ex) {
-      if (ex.response && ex.response.status === 404)
-        toast.error('This story has already been deleted');
+    if (window.confirm('Are you sure you want to delete this story')) {
+      try {
+        dispatch({
+          type: DELETE_STORY,
+          payload: id,
+        });
+        await deleteStory(id);
+        window.location.reload();
+      } catch (ex) {
+        if (ex.response && ex.response.status === 404)
+          toast.error('This story has already been deleted');
+      }
     }
   };
